requirements.txt
alembic==1.7.5
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.7.1
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
pymongo==4.0.1
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
